requirements.txt
Flask==1.1.2
jinja2==2.11.3
markupsafe==2.0.1
itsdangerous==1.1.0
werkzeug==1.0.1
cassandra-driver==3.28.0
click==8.1.3
geomet==0.2.1.post1
six==1.16.0